Create base FormattedChangeLogParser class
